Fix for Bug 692285: phase shift in patterns in long image.
The phase for a pattern mask was being held in a short, rather
than in an int. Unsurprisingly this caused problems when the
phase overflowed. Simple fix is to use an int instead.

Cluster testing shows no changes.

+++ gs/base/gsdcolor.h
@@ -321,7 +321,7 @@ struct gx_device_color_s {
 
     struct _mask {
         struct mp_ {
-            short x, y;
+            int x, y;
         } m_phase;
         gx_bitmap_id id;
         gx_color_tile *m_tile;
